get levenshteinDistance into 0..1 value. according to wikipedia: edit distance is: 	 *  "It is always at least the difference of the sizes of the two strings." 	 *  "It is at most the length of the longer string." i made min the floor and max the ceiling.

+++ java/test/org/antlr/codebuff/TestLevenshtein.java
@@ -7,7 +7,7 @@ import static org.junit.Assert.assertEquals;
 public class TestLevenshtein {
 	@Test
 	public void simple() {
-		int d = Tool.levenshteinDistance("abc", "abd");
+		float d = Tool.levenshteinDistance("abc", "abd");
 		assertEquals(1, d);
 
 		d = Tool.levenshteinDistance("kitten", "sitten");
